🎨 restructure: use esm module for helpers & 📝doc: add for loadNextRoom, fetchStoryLines, viewStoryLines
