+++ assets/storylines/demo/room-happiness.json
@@ -1,5 +1,5 @@
 [
     {
-        "text": "Hi there :3, just explore..."
+        "text": "Hi there :3, feel free to explore..."
     }
 ]
