+++ assets/storylines/navigation/happiness-room-right-door.json
@@ -1,5 +1,11 @@
 [
     {
         "text": "Sorry. You can't go back anymore..."
+    },
+    {
+        "text": "Continue with your decision. You'll be fine."
+    },
+    {
+        "text": "Aren't you?"
     }
 ]
