+++ index.pug
@@ -1,8 +1,9 @@
 doctype html
 html(lang="en")
   head
-    meta(charset="UTF-8")
-    title Merry Christmas!
+    meta(charset='UTF-8')
+    link(rel="icon", href="/assets/imgs/favicon.png", type="image/png")
+    title Xmas Cung Hỷ
     style
       | body {
       |     background-color: #1c1c1c;
