+++ src/components/door/left-door.pug
@@ -1,6 +1,7 @@
 html
   head
     link(rel='stylesheet', type='text/css', href='@@webRoot/components/door/door.css')
+    script(defer, type="module", src="@@webRoot/components/door/door.js")
   body
     .door-wrapper(id="left-door-wrapper")
       .door(id="left-door")
