+++ src/components/door/right-door.pug
@@ -1,6 +1,7 @@
 html
   head
     link(rel='stylesheet', type='text/css', href='@@webRoot/components/door/door.css')
+    script(defer, type="module", src="@@webRoot/components/door/door.js")
   body
     .door-wrapper(id="right-door-wrapper")
       .door(id="right-door")
