+++ src/components/storybox/storybox.js
@@ -8,6 +8,20 @@ document.addEventListener("DOMContentLoaded", function () {
   storyText = document.getElementById("story-content");
 });
 
+/**
+ * @description Fetches storylines from a specified path.
+ *
+ * @param {string} storyPath - The relative path (from the assets/storylines folder) to the JSON file containing the storylines.
+ *                             If null or an empty string, a default path ("demo/404-room.json") is used.
+ * @returns {Promise<Array<Object>>} A promise that resolves to an array of storyline objects.
+ *                                   Object will look like {"text": "hello world"}
+ *
+ * @example
+ * // Example usage:
+ * // Assuming assets/storylines/demo/main-room.json exists and contains an array of storylines.
+ * const storyPath = 'demo/main-room.json';
+ * const storylines = await fetchStoryLines(storyPath);
+ */
 async function fetchStoryLines(storyPath) {
   if (storyPath == null || storyPath.length == 0) {
     storyPath = "demo/404-room.json";
@@ -24,7 +38,25 @@ async function fetchStoryLines(storyPath) {
   }
 }
 
+/**
+ * @description Displays storylines in a box, allowing the player to read each line by clicking.
+ *
+ * @param {Array<Object>} storyLines - An array of storyline objects to be displayed.
+ * @param {number} [curStoryIdx=0] - The starting index for displaying storylines. Defaults to 0.
+ * 
+ * @example
+ * // Example usage:
+ * const storyLines = [
+ *   { "text": "Hi there, Im Duong from the past" },
+ *   { "text": "Saying Hello to the future me" },
+ * ];
+ * viewStoryLines(storyLines);
+ */
 function viewStoryLines(storyLines, curStoryIdx = 0) {
+  if (storyLines.length == 0) {
+    return;
+  }
+
   function viewNextStoryLine() {
     if (storyLines.length == 0 || curStoryIdx < 0) {
       storyBoxOverlay.removeEventListener("click", viewNextStoryLine);
