+++ src/rooms/main-room/index.js
@@ -1,3 +1,5 @@
+import {loadNextRoom} from '@@webRoot/components/door/door.js'
+
 // set path to intial story lines for storybox
 const intialStoryPath = "demo/main-room.json";
 const leftRoomStoryPath = "navigation/main-room-left-door.json";
@@ -9,29 +11,14 @@ document.addEventListener("DOMContentLoaded", async function () {
     viewStoryLines(intialStoryLines);
   }, 2000);
 
-  loadNextRoom(leftRoomStoryPath, 'left-door-wrapper', "../room-happiness/index.html")
-  loadNextRoom(rightRoomStoryPath, 'right-door-wrapper', "../room-suffering/index.html")
+  loadNextRoom(
+    leftRoomStoryPath,
+    "left-door-wrapper",
+    "@@webRoot/rooms/room-happiness/index.html"
+  );
+  loadNextRoom(
+    rightRoomStoryPath,
+    "right-door-wrapper",
+    "@@webRoot/rooms/room-suffering/index.html"
+  );
 });
-
-function loadNextRoom(storyPath, doorWrapperDOMElementID, roomURI) {
-  var storyLines = []
-  var isRoomEnterable = false
-  fetchStoryLines(storyPath)
-    .then((fetchedSl) => {
-      storyLines = fetchedSl;
-    })
-    .catch((e) => {
-      console.error(`Fetch story lines from ${storyPath} failed:`, e);
-    });
-
-  var doorWrapper = document.getElementById(doorWrapperDOMElementID);
-  doorWrapper.addEventListener("click", function () {
-    if (!isRoomEnterable) {
-      viewStoryLines(storyLines);
-      isRoomEnterable = true;
-      return;
-    }
-
-    window.location.href = roomURI;
-  });
-}
