+++ src/rooms/main-room/index.pug
@@ -5,7 +5,7 @@ html
     include @@webRoot/components/door/left-door.pug
     include @@webRoot/components/door/right-door.pug
     include @@webRoot/components/storybox/index.pug
-    script(defer, src="index.js")
+    script(defer, type="module", src="index.js")
     .wrapper
       ul(class="wall")
         - for (var i=0; i<15; i++)
