+++ src/rooms/room-happiness/index.pug
@@ -4,4 +4,4 @@ html
   body
     include @@webRoot/components/door/right-door.pug
     include @@webRoot/components/storybox/index.pug
-    script(defer, src="index.js")
+    script(defer, type="module",src="index.js")
