+++ src/utils/utils.js
@@ -1,0 +1,6 @@
+export function navigate(uri) {
+  if (uri.length == 0) {
+    return;
+  }
+  window.location.href = uri;
+}
